refactor(home): Rename ShortcutInfoMenu to ShortcutInfoScreen
This commit renames the `ShortcutInfoMenu` composable to `ShortcutInfoScreen`
to more accurately reflect its role as a screen component rather than
just a menu. This change improves the clarity and consistency of the
codebase.

The renaming affects the following files:
- `ShortcutInfoMenu.kt` is renamed to `ShortcutInfoScreen.kt`.
- Imports and usages of `ShortcutInfoMenu` have been updated to `ShortcutInfoScreen`
  in `GridItemPopup.kt`, `FolderGridItemPopup.kt`, and `Popup.kt`.

This change is purely a refactoring and does not alter the existing
functionality or behavior of the shortcut information display.
